+++ modules/graphics/src/cpp/GOTextInterpreterProperty.cpp
@@ -7,37 +7,18 @@
 // SPDX-License-Identifier: LGPL-3.0-or-later
 // LICENCE_BLOCK_END
 //=============================================================================
-#pragma once
-//=============================================================================
-#include "nlsGraphics_exports.h"
-#include "GraphicsObject.hpp"
+#include "GOTextInterpreterProperty.hpp"
+#include "GOPropertyValues.hpp"
 //=============================================================================
 namespace Nelson {
 //=============================================================================
-class NLSGRAPHICS_IMPEXP GOText : public GraphicsObject
+static const wchar_t* TEXT_INTERPRETER_DICT[5]
+    = { GO_PROPERTY_VALUE_TEX_STR, GO_PROPERTY_VALUE_NONE_STR, 0 };
+//=============================================================================
+GOTextInterpreterProperty::GOTextInterpreterProperty()
+    : GORestrictedStringProperty(TEXT_INTERPRETER_DICT)
 {
-private:
-    QFont fnt;
-    std::wstring text;
-    std::wstring
-    replaceSpecialCharacters(const std::wstring& text);
-
-public:
-    GOText();
-    ~GOText() override;
-    virtual std::wstring
-    getType() override;
-    virtual void
-    constructProperties();
-    virtual void
-    setupDefaults();
-    void
-    updateState() override;
-    void
-    paintMe(RenderInterface& gc) override;
-    int
-    getTextHeightInPixels();
-};
+}
 //=============================================================================
 }
 //=============================================================================