+++ CHANGELOG.md
@@ -12,16 +12,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - `DrawLater` property added to `figure` graphics object.
 - `interp1` linear interpolation 1D.
 - [736](http://github.com/Nelson-numerical-software/nelson/issues/736): `bone`, `cool`, `copper`, `hot`, `jet`, `pink`, `turbo`, `viridis`, `white` colormaps.
-- `CHANGELOG` 0.7.x family.
 - `Visible` property to `figure` graphics object.
 - [809](http://github.com/Nelson-numerical-software/nelson/issues/809): `NumberTitle` property to `figure` graphics object.
 - `AlphaMap` and `Colormap` properties added to `Axes` graphics object.
 - `LineStyleOrder` property of 'axes' used for `plot` and `plot3`.
 - `ColorOrderIndex` and `LineStyleOrderIndex` properties added to `axes` graphics object.
+- `Interpreter` property added to `text` graphics object.
+- tex special characters support for `text` graphics object.
 - `delete` for graphics objects.
 - `imread` Read image from graphics file.
 - `imwrite` Write image to graphics file.
 - `imshow` Display image.
+- `CHANGELOG` 0.7.x family.
 
 ### Changed